Add error handling callback to RPC server
Users can now register an error handling callback. This procedure is
called if an error occurs during request/response processing.

Setting the callback's stop flag to True tells the server to stop
listening for requests and terminate.

The error handling callback must be registered before instructing the
server to listen for requests.